refactor: Edit server.js filename to app.js
